organized component folder better
